requirements.txt
# MODIFIED: web search 실구현 — Tavily/SerpAPI 연동
python-dotenv>=1.0.1
tavily-python>=0.3.0
google-search-results>=2.4.2
